requirements.txt
# Core dependencies
opencv-python>=4.8.0
mediapipe>=0.10.8
numpy>=1.24.0

# 3D rendering
moderngl>=5.8.0
glfw>=2.6.0
pyrr>=0.10.3
trimesh>=4.0.0
pillow>=10.0.0

# Utilities
loguru>=0.7.0
pydantic>=2.0.0

# Optional
watchdog>=3.0.0
